requirements.txt
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
click==7.1.2
flake8==3.8.4
iniconfig==1.1.1
mccabe==0.6.1
mypy==0.800
mypy-extensions==0.4.3
numpy==1.20.1
packaging==20.9
pandas==1.2.2
pathspec==0.8.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
pytest-mock==3.5.1
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
